# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST para gestão inteligente de veículos, revisões e manutenções do CVMC (Como Vai Meu Carro).",
        "title": "CVMC API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Suporte CVMC",
            "email": "suporte@cvmc.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/cars/{id}/maintenances": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna o histórico cronológico de manutenções de um veículo",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Maintenances"
                ],
                "summary": "Listar manutenções do veículo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do veículo",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpx.SuccessEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Registra um novo registro de revisão ou manutenção para o veículo especificado",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Maintenances"
                ],
                "summary": "Cadastrar manutenção",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do veículo",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados da manutenção",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateMaintenanceRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/httpx.SuccessEnvelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
        "/api/v1/maintenances/{maintenanceID}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Atualiza dados de uma manutenção existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Maintenances"
                ],
                "summary": "Atualizar manutenção",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID da manutenção",
                        "name": "maintenanceID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados atualizados",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateMaintenanceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpx.SuccessEnvelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove o registro de manutenção",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Maintenances"
                ],
                "summary": "Excluir manutenção",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID da manutenção",
                        "name": "maintenanceID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpx.SuccessEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.CreateMaintenanceRequest": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string",
                    "example": "2026-08-23T00:00:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Óleo 0W20 Sintético + Filtro original"
                },
                "mileage": {
                    "type": "integer",
                    "example": 32000
                },
                "title": {
                    "type": "string",
                    "example": "Troca de Óleo e Filtro"
                }
            }
        },
...
        "handlers.UpdateMaintenanceRequest": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string",
                    "example": "2026-08-23T00:00:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Óleo 0W20 Sintético + Filtro original"
                },
                "mileage": {
                    "type": "integer",
                    "example": 32000
                },
                "title": {
                    "type": "string",
                    "example": "Troca de Óleo e Filtro"
                }
            }
        },
        "httpx.ErrorEnvelope": {
            "type": "object",
            "properties": {
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "httpx.SuccessEnvelope": {
            "type": "object",
            "properties": {
                "data": {},
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Digite \"Bearer \" seguido do token JWT obtido no login.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}